+++ CHANGELOG.md
@@ -22,6 +22,22 @@ The **patch** part changes is incremented if multiple releases happen the same m
 
 ### Security
 
+## [22.06] 2022-06-03
+
+### Added
+
+* certbot: add hapee (HAProxy Enterprise Edition) deploy hook
+* evolinux-base: add dir-check script
+* evolinux-base: add update-evobackup-canary script
+* mysql: add post-backup-hook to evomariabackup
+* mysql: use dir-check inside evomariabackup
+
+### Changed
+
+* docker: Allow "live-restore" to be toggled with docker_conf_live_restore
+* evocheck: upstream release 22.06
+* mysql: add "set crypt_use_gpgme=no" Mutt option, for mysqltuner
+
 ## [22.05.1] 2022-05-12
 
 ### Added